chore: take cluster logic out of the map.tsx component
